+++ src/modules/createElement.js
@@ -1,0 +1,15 @@
+const createElement = (tagName, attributes = {}) => {
+  const element = document.createElement(tagName);
+  Object.entries(attributes).forEach(([key, value]) => {
+    if (key === 'textContent') {
+      element.textContent = value;
+    } else if (key === 'onclick') {
+      element.addEventListener('click', value);
+    } else {
+      element.setAttribute(key, value);
+    }
+  });
+  return element;
+};
+
+export default createElement;
